똑장ver2/backend/pyproject.toml
[tool.poetry]
name = "ttokjang-backend"
version = "0.1.0"
description = "FastAPI Backend for Ttokjang (Smart Shopping Agent)"
authors = ["Ttokjang Team"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.128.0"
starlette = ">=0.49.1,<0.51.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
pydantic-settings = "^2.4.0"
aiosqlite = "^0.20.0"
openai = "^1.50.0"
langgraph = "^1.0.0"
langgraph-checkpoint = ">=3.0.0,<4.0.0"
langchain-openai = "^1.1.0"
langchain-core = ">=1.2.11,<2.0.0"
python-multipart = "^0.0.9"
python-dotenv = "^1.0.0"
httpx = "^0.27.0"
sqlalchemy = "^2.0.37"
alembic = "^1.14.1"
psycopg = {extras = ["binary"], version = "^3.2.9"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
httpx = "^0.27.0"
ruff = "^0.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
python_files = ["test_*.py"]
